% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get Metadata for a Specific IRPF Section}
\usage{
get_metadata(section)
}
\arguments{
\item{section}{A string representing the section name (see \code{get_sections()}).}
}
\value{
A tibble with columns \code{atributo} and \code{descricao}.
}
\description{
Returns a detailed dictionary of all attributes (columns) present in a
specific IRPF data section.
}
\examples{
get_metadata("bens_e_direitos")
}
