% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{get_sections}
\alias{get_sections}
\title{List Available IRPF Data Sections}
\usage{
get_sections()
}
\value{
A tibble with two columns: \code{secao} (the key to be used in other functions)
and \code{descricao} (a brief explanation of the data).
}
\description{
This function returns a table containing all the available data sections
that can be retrieved from the Great Numbers of Brazilian Individual Income Tax (IRPF) through this package.
}
\examples{
get_sections()
}
